requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
twilio==9.0.0
anthropic>=0.40.0
httpx==0.27.0
python-dotenv==1.0.0
